feat(UIKIT-1327,ui,baseTheme): Вынесена логика создания темы для снепшота
