+++ packages/components/package.json
@@ -31,7 +31,7 @@
     "lint:all": "eslint \"**/*.{ts,tsx}\" --fix --quiet",
     "lint:styles": "stylelint --fix",
     "lint:styles:all": "stylelint \"**/*.{ts,tsx}\" --fix",
-    "snapshot:theme": "ts-node -T ./src/theme/baseTheme/utils/generateSnapshot.ts"
+    "snapshot:theme": "ts-node -T ./src/theme/baseTheme/utils/writeSnapshot.ts"
   },
   "dependencies": {
     "@astral/icons": "*",
